semantic_kernel_example/requirements.txt
semantic-kernel
python-dotenv
